pyproject.toml
[project]
name = "sementic-search-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "ipykernel>=6.30.0",
    "numpy>=2.3.2",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
